+++ src/main/java/com/wraith/repository/CountryRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.CrudRepository;
  * Date: 24/02/13
  * Time: 16:34
  */
-public interface CountryRepository  extends CrudRepository<Country, Integer> {
+public interface CountryRepository extends CrudRepository<Country, Long> {
 
 }
