+++ src/main/java/com/wraith/repository/CurrencyRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.CrudRepository;
  * Date: 24/02/13
  * Time: 16:32
  */
-public interface CurrencyRepository extends CrudRepository<Currency, Integer> {
+public interface CurrencyRepository extends CrudRepository<Currency, Long> {
 }
