+++ src/main/java/com/wraith/repository/TransactionRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.CrudRepository;
  * Date: 24/02/13
  * Time: 15:48
  */
-public interface TransactionRepository extends CrudRepository<Transaction, Integer> {
+public interface TransactionRepository extends CrudRepository<Transaction, Long> {
 }
